refactor: move solana utils to `scaffold-utils` package
