+++ packages/base/adapters/solana/web3js/index.ts
@@ -4,12 +4,12 @@ export { SolanaWeb3JsClient } from './client.js'
 
 // -- Types -----------------------------------------------------------
 export type { Web3ModalClientOptions } from './client.js'
-export type * from './utils/scaffold/SolanaTypesUtil.js'
+export type * from '@web3modal/scaffold-utils/solana'
 export type * from '@solana/wallet-adapter-base'
 
 // -- Utils -----------------------------------------------------------
 export * from './utils/defaultConfig.js'
-export * from './utils/scaffold/SolanaStoreUtil.js'
+export * from '@web3modal/scaffold-utils/solana'
 
 // -- Constants -------------------------------------------------------
 export { solana, solanaDevnet, solanaTestnet } from './utils/chains.js'
