+++ packages/base/adapters/solana/web3js/utils/defaultConfig.ts
@@ -1,6 +1,6 @@
 import '@web3modal/polyfills'
 
-import type { Chain, Metadata, Provider, ProviderType } from './scaffold/index.js'
+import type { Chain, Metadata, Provider, ProviderType } from '@web3modal/scaffold-utils/solana'
 
 declare global {
   interface Navigator {
