--- packages/base/adapters/solana/web3js/utils/scaffold/index.ts
@@ -1,6 +1,0 @@
-export { SolConstantsUtil } from './SolanaConstantsUtil.js'
-export { SolHelpersUtil } from './SolanaHelpersUtils.js'
-export { SolStoreUtil } from './SolanaStoreUtil.js'
-
-export type { SolStoreUtilState } from './SolanaStoreUtil.js'
-export type * from './SolanaTypesUtil.js'
