+++ packages/scaffold-utils/exports/ethers.ts
@@ -1,5 +1,6 @@
-export { EthersConstantsUtil } from '../src/EthersConstantsUtil.js'
-export { EthersHelpersUtil } from '../src/EthersHelpersUtil.js'
-export { EthersStoreUtil } from '../src/EthersStoreUtil.js'
-export type { EthersStoreUtilState, Status } from '../src/EthersStoreUtil.js'
-export type * from '../src/EthersTypesUtil.js'
+export * from '../src/ethers/EthersConstantsUtil.js'
+export * from '../src/ethers/EthersHelpersUtil.js'
+export * from '../src/ethers/EthersStoreUtil.js'
+
+export type * from '../src/ethers/EthersStoreUtil.js'
+export type * from '../src/ethers/EthersTypesUtil.js'
