+++ packages/scaffold-utils/package.json
@@ -41,6 +41,8 @@
   },
   "dependencies": {
     "@coinbase/wallet-sdk": "4.0.3",
+    "@solana/web3.js": "1.95.2",
+    "@solana/wallet-adapter-base": "0.9.23",
     "@web3modal/common": "workspace:*",
     "@web3modal/core": "workspace:*",
     "@web3modal/polyfills": "workspace:*",
