+++ packages/scaffold-utils/src/ethers/EthersHelpersUtil.ts
@@ -1,6 +1,6 @@
 import type { CaipNetwork } from '@web3modal/core'
-import { ConstantsUtil } from './ConstantsUtil.js'
-import { PresetsUtil } from './PresetsUtil.js'
+import { ConstantsUtil } from '../ConstantsUtil.js'
+import { PresetsUtil } from '../PresetsUtil.js'
 import type { Chain, Provider } from './EthersTypesUtil.js'
 
 export const EthersHelpersUtil = {
