+++ packages/scaffold-utils/src/ethers/EthersTypesUtil.ts
@@ -1,6 +1,6 @@
 import type { W3mFrameProvider } from '@web3modal/wallet'
 import type { ProviderInterface } from '@coinbase/wallet-sdk'
-import type { SocialProvider } from './TypeUtil.js'
+import type { SocialProvider } from '../TypeUtil.js'
 
 export interface IEthersConfig {
   providers: ProviderType
