+++ packages/scaffold-utils/src/solana/SolanaTypesUtil.ts
